Travis CI fixes
