+++ tests/Component/Basket/BasketElementTest.php
@@ -23,19 +23,7 @@ use Sonata\Component\Product\ProductDefinition;
 use Sonata\Component\Product\ProductInterface;
 use Sonata\Component\Product\ProductManagerInterface;
 use Sonata\Component\Product\ProductProviderInterface;
-use Sonata\ProductBundle\Model\BaseProductProvider;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
-
-class ProductProviderTest extends BaseProductProvider
-{
-    /**
-     * @return string
-     */
-    public function getBaseControllerName()
-    {
-        return 'DumbTestController';
-    }
-}
 
 class BasketElementTest extends TestCase
 {
