+++ tests/Component/Delivery/Delivery.php
@@ -1,0 +1,20 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the Sonata Project package.
+ *
+ * (c) Thomas Rabaix <thomas.rabaix@sonata-project.org>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Sonata\Component\Tests\Delivery;
+
+use Sonata\ProductBundle\Entity\BaseDelivery;
+
+class Delivery extends BaseDelivery
+{
+}
