+++ tests/Component/Generator/EntityManager.php
@@ -1,0 +1,27 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the Sonata Project package.
+ *
+ * (c) Thomas Rabaix <thomas.rabaix@sonata-project.org>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Sonata\Component\Tests\Generator;
+
+use Doctrine\ORM\EntityManager as BaseEntityManager;
+
+class EntityManager extends BaseEntityManager
+{
+    public function getClassMetadata($className)
+    {
+        $obj = new \stdClass();
+        $obj->table['name'] = 'test';
+
+        return $obj;
+    }
+}
