+++ tests/Component/Generator/OrderMock.php
@@ -1,0 +1,34 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the Sonata Project package.
+ *
+ * (c) Thomas Rabaix <thomas.rabaix@sonata-project.org>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Sonata\Component\Tests\Generator;
+
+use Sonata\OrderBundle\Entity\BaseOrder;
+
+class OrderMock extends BaseOrder
+{
+    protected $id;
+
+    public function setId($id)
+    {
+        $this->id = $id;
+    }
+
+    /**
+     * @return int the order id
+     */
+    public function getId()
+    {
+        return $this->id;
+    }
+}
