+++ tests/Component/Payment/Address.php
@@ -1,0 +1,24 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the Sonata Project package.
+ *
+ * (c) Thomas Rabaix <thomas.rabaix@sonata-project.org>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Sonata\Component\Tests\Payment;
+
+use Sonata\CustomerBundle\Entity\BaseAddress;
+
+class Address extends BaseAddress
+{
+    public function getId()
+    {
+        return $this->id;
+    }
+}
