+++ tests/Component/Payment/Ogone/OgonePaymentTest.php
@@ -22,26 +22,10 @@ use Sonata\Component\Order\OrderInterface;
 use Sonata\Component\Payment\Ogone\OgonePayment;
 use Sonata\Component\Payment\TransactionInterface;
 use Sonata\Component\Product\ProductInterface;
-use Sonata\OrderBundle\Entity\BaseOrder;
 use Symfony\Bundle\FrameworkBundle\Templating\EngineInterface;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\RouterInterface;
 
-class OgonePaymentTest_Order extends BaseOrder
-{
-    public function setId($id)
-    {
-        $this->id = $id;
-    }
-
-    /**
-     * @return int the order id
-     */
-    public function getId()
-    {
-        return $this->id;
-    }
-}
 class OgonePaymentTest extends TestCase
 {
     public function testValidPayment()
