+++ tests/Component/Payment/PassPaymentTest.php
@@ -20,20 +20,9 @@ use Sonata\Component\Basket\Basket;
 use Sonata\Component\Payment\PassPayment;
 use Sonata\Component\Payment\TransactionInterface;
 use Sonata\Component\Product\ProductInterface;
-use Sonata\OrderBundle\Entity\BaseOrder;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\RouterInterface;
 
-class PassPaymentTest_Order extends BaseOrder
-{
-    /**
-     * @return int the order id
-     */
-    public function getId()
-    {
-        // TODO: Implement getId() method.
-    }
-}
 class PassPaymentTest extends TestCase
 {
     /**
