+++ tests/Component/Payment/PassPaymentTest_Order.php
@@ -1,0 +1,27 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the Sonata Project package.
+ *
+ * (c) Thomas Rabaix <thomas.rabaix@sonata-project.org>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Sonata\Component\Tests\Payment;
+
+use Sonata\OrderBundle\Entity\BaseOrder;
+
+class PassPaymentTest_Order extends BaseOrder
+{
+    /**
+     * @return int the order id
+     */
+    public function getId()
+    {
+        // TODO: Implement getId() method.
+    }
+}
