+++ tests/Component/Product/Product.php
@@ -1,0 +1,45 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the Sonata Project package.
+ *
+ * (c) Thomas Rabaix <thomas.rabaix@sonata-project.org>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Sonata\Component\Tests\Product;
+
+use Sonata\ProductBundle\Entity\BaseProduct;
+
+class Product extends BaseProduct
+{
+    public $enabled = true;
+    public $id = 1;
+    public $name = 'fake name';
+    public $price = 15;
+    public $vat = 19.6;
+
+    public function isRecurrentPayment()
+    {
+        return false;
+    }
+
+    public function getElementOptions()
+    {
+        return [];
+    }
+
+    public function getId()
+    {
+        return $this->id;
+    }
+
+    public function setId($id)
+    {
+        return $this->id = $id;
+    }
+}
